refactor and add routing between components
