--- src/components/home/Description.css
@@ -1,3 +1,0 @@
-.description {
-  line-height: 22px;
-}
