+++ src/components/home/Home.jsx
@@ -1,15 +1,14 @@
 import { useState, useEffect } from "react";
-import { BrowserRouter as Router, Route, Switch, Link } from "react-router-dom";
 import "./Home.css";
 import Side from "./Side";
-import Experience from "./Experience";
-import Description from "./Description";
 import ContactForm from "./ContactForm";
 import Footer from "./Footer";
 import { fetchData } from "../../services/api";
 import Loading from "../common/Loading";
+import Resume from "../resume/Resume";
 
 const Home = () => {
+  const [page, setPage] = useState("/");
   const [portfolio, setPortfolio] = useState(null);
   useEffect(() => {
     const getPortfolio = async () => {
@@ -25,15 +24,12 @@ const Home = () => {
 
   return portfolio ? (
     <>
-      <Side data={portfolio} />
+      <Side data={portfolio} setPage={setPage} />
       <div className="content">
-        <Description data={portfolio.description} />
-        {portfolio.experiences.map((exp) => {
-          return <Experience key={exp.company} experience={exp} />;
-        })}
+        {page === "/" && <Resume portfolio={portfolio} />}
+        {page === "/contact" && <ContactForm />}
         <Footer />
       </div>
-      <ContactForm />
     </>
   ) : (
     <div className="loading">
