--- src/components/menu/Menu.css
@@ -1,9 +1,0 @@
-.menu {
-  font-size: 14px;
-  margin-top: 50px;
-}
-.menu-item {
-  cursor: pointer;
-  list-style: none;
-  margin-top: 10px;
-}
