+++ src/components/resume/Experience.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { getMonthYear } from "../../utils/date";
 import "./Experience.css";
-import Link from "../common/Link";
 import Skill from "../common/Skill";
 
 const Experience = (props) => {
@@ -30,8 +29,27 @@ const Experience = (props) => {
           dangerouslySetInnerHTML={{ __html: props.experience.description }}
         />
         <div className="links">
-          {props.experience.links.map((lnk) => {
-            return <Link link={lnk} key={lnk.title} />;
+          {props.experience.links.map((link) => {
+            return (
+              <a
+                href={link.url}
+                key={link.url}
+                target="_blank"
+                className="underline_from_center"
+              >
+                <svg
+                  width="15"
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 20 20"
+                  fill="currentColor"
+                  aria-hidden="true"
+                >
+                  <path d="M12.232 4.232a2.5 2.5 0 013.536 3.536l-1.225 1.224a.75.75 0 001.061 1.06l1.224-1.224a4 4 0 00-5.656-5.656l-3 3a4 4 0 00.225 5.865.75.75 0 00.977-1.138 2.5 2.5 0 01-.142-3.667l3-3z"></path>
+                  <path d="M11.603 7.963a.75.75 0 00-.977 1.138 2.5 2.5 0 01.142 3.667l-3 3a2.5 2.5 0 01-3.536-3.536l1.225-1.224a.75.75 0 00-1.061-1.06l-1.224 1.224a4 4 0 105.656 5.656l3-3a4 4 0 00-.225-5.865z"></path>
+                </svg>
+                {link.title}
+              </a>
+            );
           })}
         </div>
         <div className="skills">
