+++ src/components/resume/Resume.css
@@ -1,0 +1,3 @@
+.description {
+  line-height: 1.4rem;
+}
